Update powershell-install-configure.md
I added a link to the step-by-step video I created on installing the modules and getting started using PowerShell. Not sure if this is the proper way to submit the change. I just had challenges getting going to thought I would make it easier for others. I am a SharePoint MVP if that matters at all. :)